requirements.txt
pandas
Pillow
python-dotenv
Requests
scikit_learn
streamlit
streamlit_option_menu
transformers
torch
tensorflow